vochain: upgrade to cometBFT 0.37.1
