+++ dockerfiles/testsuite/grafana/provisioning/dashboards/metrics.json
@@ -221,6 +221,101 @@
             "axisLabel": "",
             "axisPlacement": "auto",
             "barAlignment": 0,
+            "drawStyle": "bars",
+            "fillOpacity": 0,
+            "gradientMode": "none",
+            "hideFrom": {
+              "legend": false,
+              "tooltip": false,
+              "viz": false
+            },
+            "lineInterpolation": "linear",
+            "lineStyle": {
+              "fill": "solid"
+            },
+            "lineWidth": 1,
+            "pointSize": 5,
+            "scaleDistribution": {
+              "type": "linear"
+            },
+            "showPoints": "auto",
+            "spanNulls": false,
+            "stacking": {
+              "group": "A",
+              "mode": "none"
+            },
+            "thresholdsStyle": {
+              "mode": "off"
+            }
+          },
+          "mappings": [],
+          "thresholds": {
+            "mode": "absolute",
+            "steps": [
+              {
+                "color": "green",
+                "value": null
+              },
+              {
+                "color": "red",
+                "value": 80
+              }
+            ]
+          }
+        },
+        "overrides": []
+      },
+      "gridPos": {
+        "h": 8,
+        "w": 24,
+        "x": 0,
+        "y": 15
+      },
+      "id": 123132,
+      "options": {
+        "legend": {
+          "calcs": [],
+          "displayMode": "list",
+          "placement": "bottom",
+          "showLegend": true
+        },
+        "tooltip": {
+          "mode": "single",
+          "sort": "none"
+        }
+      },
+      "targets": [
+        {
+          "datasource": {
+            "type": "prometheus",
+            "uid": "PBFA97CFB590B2093"
+          },
+          "editorMode": "builder",
+          "expr": "comet_mempool_size{environment=\"test\"}",
+          "legendFormat": "{{instance}}",
+          "range": true,
+          "refId": "A"
+        }
+      ],
+      "title": "Mempool transactions",
+      "type": "timeseries"
+    },
+    {
+      "datasource": {
+        "type": "prometheus",
+        "uid": "PBFA97CFB590B2093"
+      },
+      "fieldConfig": {
+        "defaults": {
+          "color": {
+            "mode": "palette-classic"
+          },
+          "custom": {
+            "axisCenteredZero": false,
+            "axisColorMode": "text",
+            "axisLabel": "",
+            "axisPlacement": "auto",
+            "barAlignment": 0,
             "drawStyle": "line",
             "fillOpacity": 0,
             "gradientMode": "none",
@@ -266,7 +361,7 @@
         "h": 8,
         "w": 24,
         "x": 0,
-        "y": 15
+        "y": 23
       },
       "id": 123129,
       "options": {
@@ -288,13 +383,13 @@
             "uid": "PBFA97CFB590B2093"
           },
           "editorMode": "code",
-          "expr": "rate(tendermint_consensus_block_size_bytes_sum{instance=\"miner0:9090\"}[$__rate_interval])",
+          "expr": "comet_consensus_block_size_bytes{instance=\"miner0:9090\"}/1024/1024",
           "legendFormat": "{{instance}}",
           "range": true,
           "refId": "A"
         }
       ],
-      "title": "Block size",
+      "title": "Block size MiB",
       "type": "timeseries"
     },
     {
@@ -361,7 +456,7 @@
         "h": 8,
         "w": 24,
         "x": 0,
-        "y": 23
+        "y": 31
       },
       "id": 123128,
       "options": {
@@ -383,13 +478,13 @@
             "uid": "PBFA97CFB590B2093"
           },
           "editorMode": "code",
-          "expr": "60/delta(tendermint_consensus_block_interval_seconds_count{environment=\"test\"}[60s])",
+          "expr": "60/delta(comet_consensus_block_interval_seconds_count{environment=\"test\"}[60s])",
           "legendFormat": "{{instance}}",
           "range": true,
           "refId": "A"
         }
       ],
-      "title": "Panel Title",
+      "title": "Block time",
       "type": "timeseries"
     },
     {
@@ -454,7 +549,7 @@
         "h": 7,
         "w": 24,
         "x": 0,
-        "y": 31
+        "y": 39
       },
       "id": 123127,
       "options": {
@@ -476,7 +571,7 @@
             "uid": "PBFA97CFB590B2093"
           },
           "editorMode": "code",
-          "expr": "sum(delta(tendermint_consensus_total_txs{instance=\"miner0:9090\"}[1m]))",
+          "expr": "sum(delta(comet_consensus_total_txs{instance=\"miner0:9090\"}[1m]))",
           "legendFormat": "{{instance}}",
           "range": true,
           "refId": "A"
@@ -547,7 +642,7 @@
         "h": 8,
         "w": 24,
         "x": 0,
-        "y": 38
+        "y": 46
       },
       "id": 123126,
       "options": {
@@ -639,7 +734,7 @@
         "h": 8,
         "w": 24,
         "x": 0,
-        "y": 46
+        "y": 54
       },
       "id": 123125,
       "options": {
@@ -701,7 +796,7 @@
         "h": 5,
         "w": 24,
         "x": 0,
-        "y": 54
+        "y": 62
       },
       "id": 123124,
       "options": {
@@ -737,7 +832,7 @@
       "type": "bargauge"
     }
   ],
-  "refresh": false,
+  "refresh": "",
   "schemaVersion": 38,
   "style": "dark",
   "tags": [],
@@ -745,7 +840,7 @@
     "list": []
   },
   "time": {
-    "from": "now-6h",
+    "from": "now-30m",
     "to": "now"
   },
   "timepicker": {
@@ -778,6 +873,7 @@
   "timezone": "browser",
   "title": "metrics",
   "uid": "a0b46f64-ee07-487e-841c-876d5c4fa4a1",
-  "version": 2,
+  "version": 1,
   "weekStart": ""
 }
+
