+++ rpcapi/indexerhandlers.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"fmt"
 
-	tmtypes "github.com/tendermint/tendermint/types"
+	tmtypes "github.com/cometbft/cometbft/types"
 	api "go.vocdoni.io/dvote/rpctypes"
 	"go.vocdoni.io/dvote/vochain/indexer/indexertypes"
 	"go.vocdoni.io/proto/build/go/models"
