+++ vochain/tmdb_default.go
@@ -7,6 +7,6 @@
 
 package vochain
 
-import tmdb "github.com/tendermint/tm-db"
+import tmdb "github.com/cometbft/cometbft-db"
 
 const tmdbBackend = tmdb.GoLevelDBBackend
