+++ vocone/vocone.go
@@ -9,12 +9,12 @@ import (
 	"sync/atomic"
 	"time"
 
+	abcitypes "github.com/cometbft/cometbft/abci/types"
+	"github.com/cometbft/cometbft/crypto/tmhash"
+	tmprototypes "github.com/cometbft/cometbft/proto/tendermint/types"
+	tmcoretypes "github.com/cometbft/cometbft/rpc/core/types"
+	tmtypes "github.com/cometbft/cometbft/types"
 	"github.com/ethereum/go-ethereum/common"
-	abcitypes "github.com/tendermint/tendermint/abci/types"
-	"github.com/tendermint/tendermint/crypto/tmhash"
-	tmprototypes "github.com/tendermint/tendermint/proto/tendermint/types"
-	tmcoretypes "github.com/tendermint/tendermint/rpc/coretypes"
-	tmtypes "github.com/tendermint/tendermint/types"
 	"go.vocdoni.io/dvote/api"
 	"go.vocdoni.io/dvote/api/censusdb"
 	"go.vocdoni.io/dvote/config"
